requirements/pip.txt
celery==4.4.2
Django==3.0.5
djangorestframework==3.11.0
gunicorn==20.0.4
PyJWT==1.7.1
pytz==2020.1
redis==3.5.0
requests==2.23.0